💄 style: 캘린더 UI 에러 수정
